Add virtctl connection plugin, update kubevirt_vms to match the API
This adds an Ansible connection plugin that uses pexpect to login into
a Virtual Machine console and execute ansible commands there.

The dynamic inventory plugin is updated to use this connection
plugin by default for VMs. The variables needed for ssh connection
are left intact so it is still possible to request ssh access
to a VM.

Do not forget about the `ansible` annotation when starting a VM
as that is what makes the seamless login possible. See ansible
docs for details:
https://github.com/ansible/ansible/blob/devel/lib/ansible/plugins/inventory/kubevirt.py#L111

Change-Id: I563c5b9d71d252f126e1033878e510eef3c973f3
Reviewed-on: https://code.engineering.redhat.com/gerrit/167926
Continuous-Integration: rhv-devops-jenkins <rhv-devops-jenkins@redhat.com>
Tested-by: Martin Sivak <msivak@redhat.com>
Reviewed-by: Petr Horacek <phoracek@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>