Clean up stuck namespaces if deletion doesn't happen in 30s
In some setups, test namespaces get stuck in Terminating phase forever,
and the only way to fix it is by killing stuck finalizers for the
affected namespace. Specifically, 'kubernetes' finalizer is sometimes
stuck for test namespaces.

This change adds a shell script to kill finalizers, and hooks it into
namespace cleanup of the test suite. If needed, you can also manually
call the script (passing offending namespace name as an argument).

This change is a workaround and should eventually be removed once we
migrate to openshift-installer version that has the issue fixed (we are
told that the latest releases of the installer may have it fixed, though
they have other problems with e.g. libvirt so having the suite support
older versions of the installer is still helpful).

This change partially reverts 6e7fa0b01a3580914409f954509d82ff7691273b
to make run_command back public because we need it to implement the
nudge.

Change-Id: I19f32bd9e7d4dae062107569aecb2da14ae11f8d
Reviewed-on: https://code.engineering.redhat.com/gerrit/169562
Continuous-Integration: rhv-devops-jenkins <rhv-devops-jenkins@redhat.com>
Tested-by: Ihar Hrachyshka <ihrachys@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>