Allow to configure bridge CNI plugin name for test suite
CNV names the plugin as `cnv-bridge` while upstream names it as
`bridge`. This change allows to configure the different name in test
config file to pass tests on a setup that uses upstream images /
manifests for bridge CNI plugin deployment.

This patch removes methods that operate directly on YAML files. This is
to eventually move out of file based templates to regular python
dictionaries that are more flexible and don't require parsing YAML
files in the first place.

Change-Id: Ibcf0401626ababcc40791af0b0e36730ed40eb8e
Reviewed-on: https://code.engineering.redhat.com/gerrit/169650
Continuous-Integration: rhv-devops-jenkins <rhv-devops-jenkins@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>
Tested-by: Ihar Hrachyshka <ihrachys@redhat.com>