net-utility-daemonset: deploy only on 'workers'
If a node is a 'master' but *not* a 'worker', we don't want to deploy
the DaemonSet on it. It is still possible to have a single-node cluster
where the master is *also* a worker.

This will save us to filter the pods that run only on workers as well.

Change-Id: I699f0967343b696e2df14ff8fb3f459ae0f17381
Reviewed-on: https://code.engineering.redhat.com/gerrit/170782
Tested-by: Meni Yakove <myakove@redhat.com>
Continuous-Integration: rhv-devops-jenkins <rhv-devops-jenkins@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>