Document new patch verification procedure for the repo
It was not immediately clear to me that I, an author of new code for the
repo, is expected to self-verify my own patches (and not an automated
system, or a reviewer, to do so).

This patch documents this process.

Note: some believe that we should automate running tests against a
cluster on each patch proposed, but for now we'll just document the
process as it is right now.

Change-Id: Ieaf04f23e17d9f30f63cec50aa0b3e9d953c61e2
Reviewed-on: https://code.engineering.redhat.com/gerrit/170822
Tested-by: Ihar Hrachyshka <ihrachys@redhat.com>
Continuous-Integration: rhv-devops-jenkins <rhv-devops-jenkins@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>