Create host linux from 'Bridge' class and bond with nmstate
Instaed of using pod.execute to call ip command and create bridges and bonds, this
review update desiredState at NodeNetworkState for node creating the bridge and bond,
it also implements the bridge and bond deletion using it.

It install kubernets-nmstate at beginning of test and
remove it at the end similar to net-utility daemonset.

Change-Id: I5feaece5e4a81645655c052d6fcb51072ddae7f9
Signed-off-by: Quique Llorente <ellorent@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/175373
Continuous-Integration: rhv-devops-jenkins <rhv-devops-jenkins@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>
Tested-by: Meni Yakove <myakove@redhat.com>