simplify run_test_connectivity
run_test_connectivity uses its namespace argument to access the console
of its source vm. There is no reason to expose `namespace` as it can be
deduced from the source VM.

Furthermore, run_test_connectivity expects the name of the destination
VM, but does not make use of this argument.

There is more to fix in this function (it is actually two functions!)
but this would have to wait for a follow-up patch.

Change-Id: Idff4b19d6e63c5114f4ee669692484af27e86b7e
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://code.engineering.redhat.com/gerrit/176461
Continuous-Integration: rhv-devops-jenkins <rhv-devops-jenkins@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>