Basic upgrade test
Create VM from template with DataVolume and L2 network.
Migrate the VM.

add --upgrade pytest CLI option to be able to run only upgrade tests when
running with --upgrade.

Upgrade tests shouldn't run on each build since we need to test the
current build.

If we have cnv-2.1.0-34 and cnv-2.1.0-35 and we want to run our tests on
cnv-2.1.0-34 we must exclude upgrade tests from the run. if not we will
end testing at some point cnv-2.1.0-35.

Change-Id: Ia1cd5d66766350d2e70168c5d136db09223ae283