Add running vm to must gather data collection
This is needed for the 'test_resource / test_virtualmachine_resources'
must-gather test.
The test checks VirtualMachine resources, which could be present as
part of other tests, but the presence of which is not guaranteed.
This patch ensures a VirtualMachine is present when the test is run.
A further patch will actually verify that the vm is there:
https://code.engineering.redhat.com/gerrit/#/c/182357/


Change-Id: I1457d67b9c789723ed3fef0e561c722bd5735f02