VirtualMachineForTestsFromTemplate: look for hpp node in both DV and PVC
For hpp, DV and VM must reside on same node.
At the moment, VirtualMachineForTestsFromTemplate only looks for the hpp node in the DV annotations,
although it may not be there and exist only in the PVC (for example, cloned DVs will not have this annotation on them, only the underlying PVC will).
Notes:
- VirtualMachineForTests already looks in both DV and PVC (line 331)
- Fixes test CNV-3638 on HPP (Windows VM from cloned DV)

Change-Id: I89654a7d66ac5008c5b578c64edf47f480fc3656