Remove KubeMacPool tests that involve changing the MAC pool range.
Range of MAC addresses in KubeMacPool cannot be modified in the KMP
ConfigMap. The range is deternined by the Cluster Network Addons
Operator, and any attempt to modify this range in the KMP will result
in CNAO restoring the orignial range and overriding the new one.
Therefore, several related test-cases become irrelevant, and should
be removed (they were also deactivated in Polarion).

Change-Id: Ife9d946c8e5a3c55886c2f0dfb2e3f55a1977c25