Get resources from external repo.
Move resources to it's own repo to make it independent and usable to
other teams or any other need.

Make the low-level indepent from our tests, this will make sure and will
be me logic that only low-level resources will be there
and not specific tests realted functions.

And maybe in long term move it to external repo (non redhat - github)
  for all to use it.

Change-Id: I9d0e4fe3dc6a8e7cf212102a55fd440b697621ce