[Storage] Unique VM name in create_vm_from_dv
After VM deletion, VMI will still hang in for a few seconds, which caused
"409 VM is already running" in the test iterations that follow.
Failed the following tests: test_disk_image_after_import/upload/clone.
Adjusting the VM name to be unique in create_vm_from_dv to solve this.

Change-Id: I0086d2f7105710258200e2f1f2636bcef2ce9091