Check target_pod status as Running
As source_pod is no longer deleted quickly, it now remains in
Completed state, for a very long time even after the Pod is
migrated.

So currently, we Check
a) Source_pod status to change from Running to Succeeded
b) Source_pod terminate reason is Completed and
c) target_pod is in Running state

With the additional checks added, had to skip the NMO
testcase, because of the new bug found.

Change-Id: Ie611d884f3a093e4e0f863094526bd6ba5872d13
Signed-off-by: Kedar Bidarkar <kbidarka@redhat.com>