Running litmus scenario
In this patch we introduce ability to run litmus scenarion with
cnv-tests. It is enough to add or remove scenario list to run
specific scenarios implemented in cnv-chaos repo. The scenarios
are downloaded to a machine where the tests are run and later
are removed during cleanup step. It is possible to prevent removal
for development purposes by setting CNV_TESTS_CHAOS_KEEP_DATA to 1.

Change-Id: Ib979eadd318fc897d937ce926aa85433ad2d57a1
Signed-off-by: Piotr Kliczewski <pkliczew@redhat.com>