Remove smoke marker from test_bridge
When smoke tests fail, the tested build will not be moved
to osbs and will not be used in tier2 tests.
As such all componenets are tested with an older build.
To limit the potential failures and test only basic functionality
in smoke tests, network test were removed.
They will be covered in tier2.

Change-Id: I85275bd3e7c3b8c52af5b33235ee1205dc589125