[virt] Fix start_vm_if_not_running fixture
The fixture was causing matrix_updated_run_strategy to be called for
tests that might not necessarily need it (see
test_run_strategy_pause_unpause_vmi). Changed it so that the policy used
for determining expected exceptions on starting a non-running VM is
based on the VM's runStrategy, not on the one specified by the
automation matrix.

Change-Id: Ib06dcbdfb37a0201afa015f5feaa3b2c586e29e5