[Storage] W/A virtctl image-upload BZ #1911386 - File already closed
On latest virtctl, we get an error in output:
`Error when closing file","pos":"os_helper.go:34","reason":"close ./cirros.img: file already closed"`
However, the return code of the command is 0 and operation is still successfull.

As a workaround we will ignore this error when verifying stderr for now.

Change-Id: I7536ca6da271e3a747fb238517dc67de10757bc9