[Storage] Move BZ 1911386 W/A lower to run_command()
Previously the W/A was at run_virtctl_command(), where we have no knowledge of the return code.
As a result we were returning success even on operations that were not successful (rc != 0),
That just matched the W/A regex.

Now we only apply the W/A if the rc == 0 which is safer and corrects some failing TCs.

Change-Id: I057c1dc7c115f11b3d4db904bd3fa397fcf07f2d