test_migration_with_masquerade: Run migration in fixture.
The service fixture (http_service) is now called from the migration
fixture (migrate_vmb), to ensure the service is orirnially created
and verified in the original node (before the migration).

Several reasons for this change:
1. The migration should only run once. In dual-stack test, running
the migration in the test means it will run twice, which is a
waste of time.
2. If the migration runs twice, it may mean that the VM will be
migrated back to the original worker (in clusters with only 2
workers this will be the case for sure), so we might miss testing
the service and connectivity after complete migration.
3. The aim of this test is not to verfy the migration itself, but
rather that the service sustains the migration. Therefore, the more
correct way would be not to include the migration in the test itself.

Change-Id: I7c0c5e95c1c19838234638f571d3853c388992fa