Skip bond tests on bare-metal clusters.
The secondary physical NICs in our bare-metal machines are occupied
(as detailed in https://issues.redhat.com/browse/CNV-10776). This
prevents bond tests from passing, so we must temporarily disable
them, until we are assigned more NICs on BMs.
The request for more NICs (with DHCP disabled) is tracked on
https://issues.redhat.com/browse/CNV-10776.

Change-Id: Id1746a1e1cdd29fcb11b9094218ca7e215c6093f