Use short names for NNCPs
In many cases, we use the node name in the NNCP name.
Node name might be very long, e,g, when it's composed of the cluster
domain and name.
The NNCP name is used as a label in its derived NNCEs, and therefore
is limited to 63 chars max.
(https://kubernetes.io/docs/concepts/overview/working-with-objects/labels
/#syntax-and-character-set).

Change-Id: I9f74c0c6d1d9eb637b0a2ccddc3a991a697d990f