Avoid exception during pytest collect-only
Should not make calls to a func that can potentially
raise an exception in the parametrize section of a test.

Our CI is currently failing because of this call,
as it is trying to perform a `pytest collect-only` and evaluates
this potentially harmful call to get_images_server_url()
(EMEA server down atm, and hence exception raised during collection).

Change-Id: I37b77f7149f416fcbb458812499e3fb4b218e110