[Storage] Wait for scratch space storage class to update
In pyconfig_updated_default_sc we set a default storage class
if it wasn't already set in the cluster, for testing purposes.

However, we do not wait for that storage class to propagate down
to CDIConfig before testing (few seconds), resulting in None == ocs
in test_cdi_config_scratch_space_value_is_default.

Change-Id: I1adb1b84a834a200076c85bcbf00d0c19ce8f2c4