Docker: copy code instead of clone
We need to copy cnv-tests code in order to make cherry-pick work.
rigt now the flow is:
cherry-pick code to cnv-tests folder
build the container (which clone from Gerrit) so the new code from the
cherry-pick is not in the image.

run pip freeze to have more info about whet installed via pipenv

Change-Id: I546f41d2e93cea733da2f52137ee31a43153c537