[Storage] Delete dummy consumer pod with 0 grace_period
For some reason we see this pod hanging for an extra 30 seconds since 4.9,
whereas before it would disappear much quicker.

This impacts total time of tests (import starts only when it is gone),
and might explain some failures we are seeing.

Don't think we're losing anything by forcefully killing it?

Change-Id: I20918e600fd4b8ff626c4b9dd24c82fd039462be