remove old and obsolete test verifying a link
the test currently fails on latest 4.9.0 build.
the test fails because it looks a URL up that does not exist anymore
in the yaml.

the single link was replaced by three links for three OSs:
linux, mac, windows.

verifying the updated links (among other verifications) is covered
in another PR:
https://code.engineering.redhat.com/gerrit/c/cnv-tests/+/269624

Change-Id: I78d828ed46a9b1bdf598da8c78ebfe72a16f8f64