remove label_selector from ICSP get call
the ICSP.get used label selector which utilizes the hint
from the oc adm catalog mirror command:
  # Delete all ImageContentSourcePolicies generated by oc adm catalog mirror
  oc delete imagecontentsourcepolicy -l operators.openshift.org/catalog=true

in 2.5.8, the oc client was older, and its created ICSP entries
did not add the label to the ICSP entry.
thus, the code fetching the ICSP entries for deletion did not
get it, and the ICSP create command that followed, failed
creating an ICSP entry with the same name ("iib").

upgrade paths may vary, so that ICSP entries are created by older
oc binaries that does not add the particular label to the ICSP entry and remain.
since it is not possible to anticipate the upgrade path, it is better to remove
the label filter.

note:
label selector was added to be more surgical following a concern raised
in the merged PR:
https://code.engineering.redhat.com/gerrit/c/cnv-tests/+/284633/1//COMMIT_MSG#7

Change-Id: Ifd1c3dac821d5a086d7a67697a26032a209c16e7