Fix the way SR-IOV-capable nodes are filtered
All the nodes in an allegedly SRIOV cluster were actively labeled as
sriov-capable by the test, while it's the role of SR-IOV operator to
label only the nodes which actually support SR-IOV (i.e. have SR-IOV
supporting NICs).
This change leads to a corrrect flow, and also allows to run SR-IOV
tests on a cluster where not all of the nodes support SR-IOV.
Additonally - fix test_sriov_interfaces_post_reboot, which failed due to
not waiting for the rebooted VM's interfaces .

Change-Id: I66da7d8b581405cf5e1df7063e5bad87d514050b