Add a sampler for port_type in nns for bridge creation
From time to time the NNS interfaces take longer to get updated
with the new port.
This solves the linux bridge creation. OVS bridges bond tests don't
alwayes work and will be addresses in another PR.
This PR break the capture tests (located in test_nmpolicy_capture.py).
This will be addressed in a different PR.

Change-Id: Ia830702e49849c960ba8d7bba5d9105df2cfdeb4