Use 'test' domain
https://www.rfc-editor.org/rfc/rfc6761.html#section-6.2

In the case of using 'test' as a top level domain (TLD),
we are provided with some basic guarantees that our DNS query
will not be transmitted beyond the first DNS server we query.
There is no authoritative holder of the 'test' domain,
so any resolving DNS will come up empty.

Change-Id: I4596e909cb1839b2e6efc2873bb65961c8ade6cc