Add sampler to test_linux_bridge_before_upgrade
When test_linux_bridge_before_upgrade is ran as part of the test suite
it runs after test_bridge_marker_before_upgrade. Both tests share the
same resources, and it takes about a second for the vmi's ip address to
be available for the ping assert in test_linux_bridge_before_upgrade.

Change-Id: If11ccdc19222fe8d537c80c5bf9de019f33c8313