[Storage] Handle timeout test_clone_from_block_to_fs_using_dv_template
CDI resizes the disk to the users request, which means that even though cirros
doesn't require 3Gi, we'd still pump it to that, and then would have to
HTTP upload it to the target cloned PVC which not very efficient time-wise.
(This test does naive host-assisted cloning, so has to stream the entire size).

Change-Id: I8c7057969025d14b9cf8a6492d9370ac600761a6
Reviewed-on: https://code.engineering.redhat.com/gerrit/c/cnv-tests/+/412980
Continuous-Integration: CNV-QE Gerrit <no-reply-cnv-qe-devops@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>
Reviewed-by: Dalia Frank <dafrank@redhat.com>
Reviewed-by: Debarati Basu-Nag <dbasunag@redhat.com>
Tested-by: Dalia Frank <dafrank@redhat.com>