[Storage] Fix flaky test_successful_import_image
The test has 8 parameters, we randomly fail to pod.exec in one of them
with error message that container is not started or running.

Usually it takes 0.00015 seconds for a container to be ready to exec,
but sometimes it's 12 - 15 seconds.
When we name the pods differently, we don't hit this issue.

Change-Id: Ib256b447c2330adb4dc39117c630865ef3f4c46f
Reviewed-on: https://code.engineering.redhat.com/gerrit/c/cnv-tests/+/420912
Continuous-Integration: CNV-QE Gerrit <no-reply-cnv-qe-devops@redhat.com>
Tested-by: Jenia Peimer <jpeimer@redhat.com>
Reviewed-by: Debarati Basu-Nag <dbasunag@redhat.com>
Reviewed-by: Meni Yakove <myakove@redhat.com>
Reviewed-by: Sarah Bennert <sbennert@redhat.com>
Reviewed-by: Satyajit Bulage <sbulage@redhat.com>
Reviewed-by: Yan Du <yadu@redhat.com>