Fixture to get common CPU models list
This patch is the forward port of the patch[1] in CNV-4.10 automation
branch. This patch introduces a new fixture 'nodes_common_cpu_model_list'
which gets the list of common CPU model across the given set of nodes.
Also the older logic to get the latest CPU model by comparing with the
static list of recent CPU family is removed.

[1] - https://code.engineering.redhat.com/gerrit/c/cnv-tests/+/420408

Change-Id: I4c0c047bd3f0cb13abd4bbbcd4bd35bcf2cb0f95
Reviewed-on: https://code.engineering.redhat.com/gerrit/c/cnv-tests/+/422829
Continuous-Integration: CNV-QE Gerrit <no-reply-cnv-qe-devops@redhat.com>
Reviewed-by: Debarati Basu-Nag <dbasunag@redhat.com>
Reviewed-by: Satyajit Bulage <sbulage@redhat.com>
Reviewed-by: Sarah Bennert <sbennert@redhat.com>
Reviewed-by: Geetika Kapoor <gkapoor@redhat.com>
Tested-by: Satheesaran Sundaramoorthi <sasundar@redhat.com>